second iteration

+++ DatingAppLibrary/UserProfile.cs
@@ -18,8 +18,8 @@ namespace DatingAppLibrary
         public string Phone { get; set; }
         public string Occupation { get; set; }
         public int Age { get; set; }
-        public double Height { get; set; }
-        public double Weight { get; set; }
+        public int Height { get; set; }
+        public int Weight { get; set; }
         public string ProfilePhotoURL { get; set; }
         public string Interests { get; set; }
         public string Likes { get; set; }
